proper wrapping of tool in context
